🔧 jest 전역 import 설정 (#32)
