--- jest.config.ts
@@ -1,14 +1,0 @@
-import { type JestConfigWithTsJest, createDefaultPreset } from 'ts-jest';
-
-const defaultPreset = createDefaultPreset();
-
-const jestConfig: JestConfigWithTsJest = {
-  testEnvironment: 'jsdom', // web app
-  setupFilesAfterEnv: ['<rootDir>/node_modules/@testing-library/jest-dom'],
-  moduleNameMapper: {
-    '^@src/(.*)$': '<rootDir>/src/$1',
-  },
-  ...defaultPreset,
-};
-
-export default jestConfig;
